chore: rename solution files
