+++ dynamic_programming/quickest_concat/quickest_concat_v2.py
@@ -1,18 +1,32 @@
-from typing import Union
-def quickest_concat(s:str, words:list[str]) -> int:
-    result = _quickest_concat(s, words)
+def quickest_concat(s, words):
+    result = _quickest_concat(s, words, {})
     if result == float('inf'):
         return -1
-    return result
+    else:
+        return result
 
-def _quickest_concat(s:str, words:list[str]) -> Union[float,int]:
-    if len(s) == 0:
+def _quickest_concat(s, words, memo):
+    if s in memo:
+        return memo[s]
+    
+    if s == '':
         return 0
+    
     min_words = float('inf')
     for word in words:
         if s.startswith(word):
             suffix = s[len(word):]
-            num_words = 1 + _quickest_concat(suffix, words)
-            min_words = min(num_words, min_words)
-            
-    return min_words            
+            attempt = 1 + _quickest_concat(suffix, words, memo)
+            min_words = min(attempt, min_words)
+    
+    memo[s] = min_words
+    return min_words
+
+"""
+
+s = length of string
+w = # of words
+Time: ~O(sw)
+Space: O(s)
+
+"""
